network_scanner/requirements.txt
flask==3.0.0
waitress==3.0.0
python-nmap==0.7.1
netifaces==0.11.0
requests==2.31.0
mac-vendor-lookup==0.1.12
